refactor: remove setup runtime onboarding flow
Removal manifesto:

- Removed hidden setup --backend runtime warmup, setup --no-open-ide, interactive terminal onboarding prompts, SETUP_RUNTIME orchestration/output, onboarding config state, and the dialoguer dependency.

- Authority: ADR 0006 separates repository setup from runtime lifecycle and lists setup extension points without backend/onboarding flags; ADR 0007 and ADR 0010 put macOS onboarding on install.sh plus plugin-owned workspace setup.

- Not V1 because setup should install only the skill and managed guidance region; runtime startup belongs under kast developer runtime and macOS workspace preparation belongs to the plugin.

- Replaced by kast setup for repository guidance, kast developer runtime up/status for backend lifecycle, and install.sh plus kast developer machine plugin for macOS machine onboarding.

- Removed ambiguity around whether setup is also an IDE prompt, runtime warmer, or terminal onboarding flow.

- Validation: cargo fmt --manifest-path cli-rs/Cargo.toml --all -- --check; cargo test --manifest-path cli-rs/Cargo.toml --locked --test cli_core_smoke --test agent_setup_smoke --test repo_resource_smoke; cargo test --manifest-path cli-rs/Cargo.toml --locked project_open; cargo clippy --manifest-path cli-rs/Cargo.toml --locked --all-targets --all-features -- -D warnings; .github/scripts/test-terminal-command-contract.sh; .github/scripts/test-docs-content-contract.sh; .github/scripts/test-release-workflow-contract.sh; rg stale setup-runtime/onboarding symbols; git diff --check.
