+++ cli-rs/Cargo.toml
@@ -27,7 +27,6 @@ uuid = { version = "1.18", features = ["v4"] }
 zip = { version = "6.0", default-features = false, features = ["deflate"] }
 jsonschema = { version = "0.46.5", default-features = false }
 indicatif = "0.18.4"
-dialoguer = { version = "0.12.0", default-features = false }
 
 [build-dependencies]
 serde_json = "1.0"
