+++ cli-rs/src/config/tests.rs
@@ -535,8 +535,6 @@ dynamicOutput = false
         assert_eq!(config.project_open.profile, ProjectOpenProfile::CopilotLsp);
         assert!(config.project_open.auto_exclude_git);
         assert!(config.project_open.gradle_load_enabled);
-        assert!(!config.onboarding.setup_completed);
-        assert!(config.can_run_setup_onboarding());
     }
 
     #[test]
@@ -561,45 +559,6 @@ gradleLoadEnabled = false
         assert_eq!(config.project_open.profile, ProjectOpenProfile::CopilotLsp);
         assert!(!config.project_open.auto_exclude_git);
         assert!(!config.project_open.gradle_load_enabled);
-        assert!(!config.can_run_setup_onboarding());
-    }
-
-    #[test]
-    fn parses_setup_onboarding_state() {
-        let temp = tempfile::tempdir().unwrap();
-        let config_file = temp.path().join("config.toml");
-        fs::write(
-            &config_file,
-            r#"[onboarding]
-setupCompleted = true
-"#,
-        )
-        .unwrap();
-
-        let mut config = KastConfig::defaults();
-        config.apply(read_partial_config(&config_file).unwrap());
-
-        assert!(config.onboarding.setup_completed);
-        assert!(!config.can_run_setup_onboarding());
-    }
-
-    #[test]
-    fn ignores_legacy_agent_up_onboarding_state() {
-        let temp = tempfile::tempdir().unwrap();
-        let config_file = temp.path().join("config.toml");
-        fs::write(
-            &config_file,
-            r#"[onboarding]
-agentUpCompleted = true
-"#,
-        )
-        .unwrap();
-
-        let mut config = KastConfig::defaults();
-        config.apply(read_partial_config(&config_file).unwrap());
-
-        assert!(!config.onboarding.setup_completed);
-        assert!(config.can_run_setup_onboarding());
     }
 
     #[test]
